Update patch set 2
Patch Set 2: Code-Review+1

PS2: rebased

Namagi: you can edit the commit message directly via the gerrit web interface (the small edit icon right to the Permalink at the top right)

Though the dalvik.vm.dexopt-data-only default has been changed to 1 for a reason and it's been recommended to not change it, I still think that the 756MB cache on xt897 can handle it easily. 4.6GB of data on xt897 is pretty limited, so I understand that the few tens of megabytes freed by this could be handy in some situations.

Well, the most storage space waste happens in system partition - currently, with CM11 installed, we have 950MB free there on xt897...

Patch-set: 2