Update patch set 3
Patch Set 3:

PS3: Commit message updated.

Pavel: thanks. However, pushing directly worked this time. Go figure. The problem with previous attempts was that I could not update it in any way, not just the message. Perhaps it was some temporary server error as the local branch is unchanged (apart from rebase. hm...)


System is indeed wasting a lot of usefull space. Without multibooting or reformating which is not possible (i think) it is just standing there. Perhaps isolating part of it into some kind of sandboxed environment and using it as an extension to /data. Not sure how, though, or if it would be worth the trouble.


What were the original reasons to keep dex in data only?

Patch-set: 3
Reviewer: namagi <6606@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0