Update patch set 1
Patch Set 1: Code-Review-1

I will test room service again but it was working before. This is not the right way though. It is done this way so that you don't end up with a zip named cm-11-20140528-NIGHTLY-msm8960.zip as then people would be very confused as to what device it was for.

Patch-set: 1
Label: Code-Review=-1