Update patch set 1
Patch Set 1:

Michael: How is it supposed to work? Should I join the translator teams for languages that I don't understand at all (as e.g. Arabic) to be able to fix obvious errors as QWERTY used in place of QWERTZ (as in the Arabic and German translations in this case)?

Isn't there a mechanism to propagate the changes submitted via gerrit back to translate.cyanogenmod.org?

This commit aimed at fixing https://jira.cyanogenmod.org/browse/CYAN-4386 , besides the ordering issue.

Patch-set: 1