Update patch set 1
Patch Set 1: Code-Review-1

Stock snd_soc_msm_2x?

Please try to describe the issue in more detail. I'm not sure what 'mic issue' here means. Knowing what exactly is supposed to be fixed is necessary for verification.

Please try to identify which exact parts of this change actually fix the issue.

This commit as it is, among other things:

1) - (re)introduces a lot of TAB vs. space noise cleaned-up before (so much that it's very hard to see the actual functional changes)

2) - decreases the ear piece volume (is it intended?)

3) - removes the "Dock" device use case => kills the support for analog audio output via USB in CM (used by Motorola Car Docks)

4) - re-introduces the camcorder audio use cases usable only by stock audio HAL, not supported by CAF audio HAL used in CM

5) - re-introduces the dock audio use cases usable only by stock audio HAL, not supported by CAF audio HAL used in CM

So, as it is now, it surely can't be merged, at least because of point 1) and 3).

I'd be really grateful if you could isolate the changes necessary for the fix instead of blind replace of the whole snd_soc_msm_2x with the stock copy. Thanks.

Patch-set: 1
Reviewer: Pavel Kucera <129@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1